+++ CsharpBetaExecutionTests/SnippetExecutionBetaTests.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.Identity.Client;
+using MsGraphSDKSnippetsCompiler;
 using MsGraphSDKSnippetsCompiler.Models;
 using NUnit.Framework;
 using TestsCommon;
@@ -9,6 +11,28 @@ namespace CsharpBetaExecutionTests
     [TestFixture]
     public class SnippetExecutionBetaTests
     {
+        public static IConfidentialClientApplication confidentialClientApp;
+        public static IPublicClientApplication publicClientApp;
+
+        [OneTimeSetUp]
+        public void Init()
+        {
+            var config = AppSettings.Config();
+            var clientId = config.GetNonEmptyValue("ClientID");
+            var authority = config.GetNonEmptyValue("Authority");
+            var username = config.GetNonEmptyValue("Username");
+            var password = config.GetNonEmptyValue("Password");
+            // application permissions
+            var tenantId = config.GetNonEmptyValue("TenantID");
+            var clientSecret = config.GetNonEmptyValue("ClientSecret");
+            publicClientApp = PublicClientApplicationBuilder.Create(clientId).WithAuthority(authority).Build();
+            confidentialClientApp = ConfidentialClientApplicationBuilder
+                .Create(clientId)
+                .WithTenantId(tenantId)
+                .WithClientSecret(clientSecret)
+                .Build();
+
+        }
         /// <summary>
         /// Gets TestCaseData for Beta
         /// TestCaseData contains snippet file name, version and test case name
@@ -28,10 +52,10 @@ namespace CsharpBetaExecutionTests
         /// <param name="docsLink">documentation page where the snippet is shown</param>
         /// <param name="version">Docs version (e.g. V1, Beta)</param>
         [Test]
-        [TestCaseSource(typeof(SnippetExecutionBetaTests), nameof(TestDataBeta))]
+        [RetryTestCaseSource(typeof(SnippetExecutionBetaTests), nameof(TestDataBeta), MaxTries = 3)]
         public async Task Test(ExecutionTestData testData)
         {
-            await CSharpTestRunner.Execute(testData);
+            await CSharpTestRunner.Execute(testData,publicClientApp,confidentialClientApp);
         }
     }
 }
